requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
supabase==2.0.0
httpx==0.24.1
python-multipart==0.0.20
python-dotenv==1.0.0
pydantic==2.5.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0